+++ ostap/fitting/tests/test_fitting_toys.py
@@ -42,7 +42,23 @@ fit_gauss = Models.Gauss_pdf ( 'FG' , xvar = mass )
 gen_gauss.mean  = 0.4
 gen_gauss.sigma = 0.1
 
+# ==============================================================================
+## Perform toy-study for possible fit bias and correct uncertainty evaluation
+#  - generate <code>nToys</code> pseudoexperiments with some PDF <code>pdf</code>
+#  - fit teach experiment with the same PDF
+#  - store  fit results
+#  - calculate staistics of pulls
+#  - fill distributions for fit results
+#  - fill distribution of pulls 
 def test_toys ( ) :
+    """Perform toys-study for possible fit bias and correct uncertainty evaluation
+    - generate `nToys` pseudoexperiments with some PDF `pdf`
+    - fit teach experiment with the same PDF
+    - store  fit results
+    - calculate statistics of pulls
+    - fill distributions of fit results
+    - fill distributions of pulls 
+    """
     
     results , stats = Toys.make_toys  (
         pdf         = gen_gauss ,
@@ -72,8 +88,18 @@ def test_toys ( ) :
         time.sleep ( 1 )
 
 # =============================================================================
+## Perform toy-study for possible fit bias and correct uncertainty evaluation
+#  - generate <code>nToys</code> pseudoexperiments with some PDF <code>gen_pdf</code>
+#  - fit teach experiment with the PDF <code>fit_pdf</code>
+#  - store  fit results
+#  - fill distributions for fit results
 def test_toys2 ( ) :
-                
+    """Perform toys-study for possible fit bias and correct uncertainty evaluation
+    - generate `nToys` pseudoexperiments with some PDF `gen_pdf`
+    - fit teach experiment with the PDF `fit_pdf`
+    - store  fit results
+    - fill distributions of fit results
+    """    
     results , stats = Toys.make_toys2 (
         gen_pdf     = gen_gauss ,
         fit_pdf     = fit_gauss ,
@@ -104,11 +130,68 @@ def test_toys2 ( ) :
         time.sleep ( 1 )
 
 # =============================================================================
+## Perform toy-study for significance of the signal 
+#  - generate <code>nToys</code> pseudoexperiments using background-only hypothesis 
+#  - fit teach experiment with "signal+background" hypothesis
+#  - store  fit results
+#  - fill distributions for fit results
+def test_significance_toys ( ) :
+    """Perform toy-study for significance of the signal 
+    - generate `nToys` pseudoexperiments using background-only hypothesis 
+    - fit each experiment with signal+background hypothesis
+    - store  fit results
+    - fill distributions for fit results
+    """
+    
+    ## only background hypothesis
+    bkg_only = Models.Bkg_pdf    ( "BKG" , xvar =  mass , power = 0 , tau = 0      )
+
+    
+    signal   = Models.Gauss_pdf  ( 'S'   , xvar = mass , mean = 0.5 , sigma = 0.1 )
+
+    signal.mean .fix ( 0.4 ) 
+    signal.sigma.fix ( 0.1 ) 
+
+
+    ## signal + background hypothesis
+    model    = Models.Fit1D      ( signal = signal , background = 1 )
+    model.background.tau.fix ( 0 )
+    
+    results , stats = Toys.make_toys2 (
+        gen_pdf     = bkg_only  ,
+        fit_pdf     = model     ,
+        nToys       = 1000      ,
+        data        = [ mass ]  , 
+        gen_config  = { 'nEvents'  : 100 , 'sample'   : True } ,
+        fit_config  = { 'silent'   : True } ,
+        gen_pars    = { 'tau_BKG'  : 0.   } , ## initial values for generation 
+        fit_pars    = { 'B' : 100         ,
+                        'S' : 10          ,
+                        'phi0_Bkg_S': 0.0 } , ## initial fit values for parameters 
+        silent      = True , 
+        progress    = True )
+
+    for p in stats :
+        logger.info (  "Toys: %-20s : %s" % (  p, stats [ p ] ) )
+
+    h_S      = ROOT.TH1F ( hID() , '#S' , 60 ,  0 , 60 )
+    
+    for r in results ['S'  ] : h_S .Fill ( r )
+    
+    for h in ( h_S ,  ) :
+        
+        h.draw()
+        logger.info ( "%s  :\n%s"  % ( h.GetTitle() , h.dump ( 30 , 10 ) ) )
+        time.sleep  ( 1 )
+
+            
+# =============================================================================
 if '__main__' == __name__ :
 
     test_toys  () 
     test_toys2 () 
-
+    test_significance_toys ( ) 
+    
 
 # =============================================================================
 ##                                                                      The END